🧪 consolidate duplicated signals adapter smoke tests
